backend/MCQ_GENERATOR/requirements.txt
# Required packages for the project

# OpenAI package for AI-related functionalities
openai

# Google Generative AI for generative AI functionalities
google-generativeai

# python-dotenv for managing environment variables from a .env file
python-dotenv

# to fetch youtube transcripts from youtube urls
youtube-transcript-api

# requests for making HTTP requests to MongoDB API
pymongo

# for ObjectId in MongoDB
bson

# # Langchain for blockchain-based language applications
# langchain

# # Streamlit for building interactive web apps with Python
# streamlit

# langchain_community
# langchain-openai

# # To install the local package in the environment,
# # you can either use 'python setup.py install'
# # or just add '-e .' to install in editable mode. 
# # -e .


# # installing the requirements.txt generates a 'mcq_generator.egg-info' with info regarding the installed stuff in the enviorment
